feat:#5 체크리스트 컴포넌트 구조 변경(메인 이동)
